backend-drf/requirements.txt
absl-py==2.4.0
asgiref==3.11.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.61.1
frozendict==2.4.7
h5py==3.15.1
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib
mdurl==0.1.2
multitasking==0.0.12
namex==0.1.0
numpy<2
optree==0.18.0
packaging==26.0
pandas==3.0.0
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
pycparser==3.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-decouple
pytz==2025.2
requests==2.32.5
rich==14.3.1
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
yfinance
tensorflow==2.15.0
gunicorn
whitenoise